requirements.txt
# Core packages
requests
numpy
python-dotenv
httpx

# FastAPI and related
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
jinja2

# Pinecone
pinecone-client>=2.0.0
langchain-pinecone

# NLP / Embeddings
scikit-learn
scipy

# LangChain related
langchain
langchain-core
langchain-google-genai
langchain-openai
openai

# Optional caching / utilities
cachetools



langgraph
langchain-groq
chromadb
